backend/pyproject.toml
[project]
name = "bring-the-village-back"
version = "0.1.0"
description = "Community platform for single parents to match and support each other"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.15",
    "django-cors-headers>=4.3",
    "gunicorn>=22.0",
    "anthropic>=0.40",
    "dj-database-url>=2.1",
    "psycopg2-binary>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.ruff]
target-version = "py311"
line-length = 120
exclude = ["*/migrations/*"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "village.settings"
python_files = ["tests.py", "test_*.py"]
addopts = "--tb=short -q"
